Donaciones.

+++ app/Livewire/Crud/Donaciones/EliminarDonacionesModal.php
@@ -13,20 +13,20 @@ class EliminarDonacionesModal extends Component
     // Método para eliminar la donación
     public function deleteItem()
     {
-        $this->item->delete();  // Eliminar la donación
-        $this->dispatch('cerrar-modal');  // Emitir evento para cerrar el modal
-        $this->dispatch('item-deleted');  // Emitir evento para notificar que la donación fue eliminada
+        // Validar si el item existe antes de eliminar
+        if ($this->item) {
+            $this->item->delete();  // Eliminar la donación
+            $this->dispatch('cerrar-modal');  // Emitir evento para cerrar el modal
+            $this->dispatch('item-deleted');  // Emitir evento para notificar que la donación fue eliminada
+        }
     }
 
-    // Método para inicializar los parámetros (en caso de que necesites lógica adicional)
+    // Método para inicializar los parámetros
     public function mount($parameters)
     {
         $this->item = $parameters['item'];  // La donación a eliminar
         $this->idModal = $parameters['idModal'];  // ID del modal
     }
-
-    // Método vacío para posibles inicializaciones adicionales
-    public function initInfo() {}
 
     // Renderizar la vista del modal
     public function render()
